requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==4.2.7
djangorestframework==3.14.0

# ── API documentation ─────────────────────────────────────────────────────────
drf-yasg==1.21.7
setuptools==69.0.2

# ── Environment / config ──────────────────────────────────────────────────────
django-environ==0.11.2

# ── CORS ──────────────────────────────────────────────────────────────────────
django-cors-headers==4.3.1

# ── Static files (production) ─────────────────────────────────────────────────
whitenoise==6.6.0

# ── Production WSGI server ────────────────────────────────────────────────────
gunicorn==21.2.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-django==4.7.0
